requirements.txt
pyzipper
PyYAML
imap-tools